+++ internal/crossdock/server/yarpc/gauntlet.go
@@ -29,52 +29,64 @@ import (
 	"go.uber.org/yarpc/internal/crossdock/thrift/gauntlet"
 )
 
-func resMetaFromReqMeta(reqMeta yarpc.ReqMeta) yarpc.ResMeta {
-	return yarpc.NewResMeta().Headers(reqMeta.Headers())
+func copyRequestHeaders(ctx context.Context) {
+	call := yarpc.CallFromContext(ctx)
+	for _, k := range call.HeaderNames() {
+		if err := call.WriteResponseHeader(k, call.Header(k)); err != nil {
+			panic(err)
+		}
+	}
 }
 
 // thriftTest implements the ThriftTest service.
 type thriftTest struct{}
 
-func (thriftTest) TestBinary(ctx context.Context, reqMeta yarpc.ReqMeta, thing []byte) ([]byte, yarpc.ResMeta, error) {
-	return thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestBinary(ctx context.Context, thing []byte) ([]byte, error) {
+	copyRequestHeaders(ctx)
+	return thing, nil
 }
 
-func (thriftTest) TestByte(ctx context.Context, reqMeta yarpc.ReqMeta, thing *int8) (int8, yarpc.ResMeta, error) {
-	return *thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestByte(ctx context.Context, thing *int8) (int8, error) {
+	copyRequestHeaders(ctx)
+	return *thing, nil
 }
 
-func (thriftTest) TestDouble(ctx context.Context, reqMeta yarpc.ReqMeta, thing *float64) (float64, yarpc.ResMeta, error) {
-	return *thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestDouble(ctx context.Context, thing *float64) (float64, error) {
+	copyRequestHeaders(ctx)
+	return *thing, nil
 }
 
-func (thriftTest) TestEnum(ctx context.Context, reqMeta yarpc.ReqMeta, thing *gauntlet.Numberz) (gauntlet.Numberz, yarpc.ResMeta, error) {
-	return *thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestEnum(ctx context.Context, thing *gauntlet.Numberz) (gauntlet.Numberz, error) {
+	copyRequestHeaders(ctx)
+	return *thing, nil
 }
 
-func (thriftTest) TestException(ctx context.Context, reqMeta yarpc.ReqMeta, arg *string) (yarpc.ResMeta, error) {
+func (thriftTest) TestException(ctx context.Context, arg *string) error {
 	switch *arg {
 	case "Xception":
 		code := int32(1001)
-		return nil, &gauntlet.Xception{ErrorCode: &code, Message: arg}
+		return &gauntlet.Xception{ErrorCode: &code, Message: arg}
 	case "TException":
 		// TODO raise TException once we support it. Meanwhile, return an
 		// unexpected exception.
-		return nil, errors.New("great sadness")
+		return errors.New("great sadness")
 	default:
-		return resMetaFromReqMeta(reqMeta), nil
+		copyRequestHeaders(ctx)
+		return nil
 	}
 }
 
-func (thriftTest) TestI32(ctx context.Context, reqMeta yarpc.ReqMeta, thing *int32) (int32, yarpc.ResMeta, error) {
-	return *thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestI32(ctx context.Context, thing *int32) (int32, error) {
+	copyRequestHeaders(ctx)
+	return *thing, nil
 }
 
-func (thriftTest) TestI64(ctx context.Context, reqMeta yarpc.ReqMeta, thing *int64) (int64, yarpc.ResMeta, error) {
-	return *thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestI64(ctx context.Context, thing *int64) (int64, error) {
+	copyRequestHeaders(ctx)
+	return *thing, nil
 }
 
-func (thriftTest) TestInsanity(ctx context.Context, reqMeta yarpc.ReqMeta, argument *gauntlet.Insanity) (map[gauntlet.UserId]map[gauntlet.Numberz]*gauntlet.Insanity, yarpc.ResMeta, error) {
+func (thriftTest) TestInsanity(ctx context.Context, argument *gauntlet.Insanity) (map[gauntlet.UserId]map[gauntlet.Numberz]*gauntlet.Insanity, error) {
 	result := map[gauntlet.UserId]map[gauntlet.Numberz]*gauntlet.Insanity{
 		1: {
 			gauntlet.NumberzTwo:   argument,
@@ -84,18 +96,21 @@ func (thriftTest) TestInsanity(ctx context.Context, reqMeta yarpc.ReqMeta, argum
 			gauntlet.NumberzSix: &gauntlet.Insanity{},
 		},
 	}
-	return result, resMetaFromReqMeta(reqMeta), nil
+	copyRequestHeaders(ctx)
+	return result, nil
 }
 
-func (thriftTest) TestList(ctx context.Context, reqMeta yarpc.ReqMeta, thing []int32) ([]int32, yarpc.ResMeta, error) {
-	return thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestList(ctx context.Context, thing []int32) ([]int32, error) {
+	copyRequestHeaders(ctx)
+	return thing, nil
 }
 
-func (thriftTest) TestMap(ctx context.Context, reqMeta yarpc.ReqMeta, thing map[int32]int32) (map[int32]int32, yarpc.ResMeta, error) {
-	return thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestMap(ctx context.Context, thing map[int32]int32) (map[int32]int32, error) {
+	copyRequestHeaders(ctx)
+	return thing, nil
 }
 
-func (thriftTest) TestMapMap(ctx context.Context, reqMeta yarpc.ReqMeta, hello *int32) (map[int32]map[int32]int32, yarpc.ResMeta, error) {
+func (thriftTest) TestMapMap(ctx context.Context, hello *int32) (map[int32]map[int32]int32, error) {
 	result := map[int32]map[int32]int32{
 		-4: {
 			-4: -4,
@@ -110,10 +125,11 @@ func (thriftTest) TestMapMap(ctx context.Context, reqMeta yarpc.ReqMeta, hello *
 			4: 4,
 		},
 	}
-	return result, resMetaFromReqMeta(reqMeta), nil
+	copyRequestHeaders(ctx)
+	return result, nil
 }
 
-func (thriftTest) TestMulti(ctx context.Context, reqMeta yarpc.ReqMeta, arg0 *int8, arg1 *int32, arg2 *int64, arg3 map[int16]string, arg4 *gauntlet.Numberz, arg5 *gauntlet.UserId) (*gauntlet.Xtruct, yarpc.ResMeta, error) {
+func (thriftTest) TestMulti(ctx context.Context, arg0 *int8, arg1 *int32, arg2 *int64, arg3 map[int16]string, arg4 *gauntlet.Numberz, arg5 *gauntlet.UserId) (*gauntlet.Xtruct, error) {
 	hello := "Hello2"
 	result := &gauntlet.Xtruct{
 		StringThing: &hello,
@@ -121,56 +137,65 @@ func (thriftTest) TestMulti(ctx context.Context, reqMeta yarpc.ReqMeta, arg0 *in
 		I32Thing:    arg1,
 		I64Thing:    arg2,
 	}
-	return result, resMetaFromReqMeta(reqMeta), nil
+	copyRequestHeaders(ctx)
+	return result, nil
 }
 
-func (thriftTest) TestMultiException(ctx context.Context, reqMeta yarpc.ReqMeta, arg0 *string, arg1 *string) (*gauntlet.Xtruct, yarpc.ResMeta, error) {
+func (thriftTest) TestMultiException(ctx context.Context, arg0 *string, arg1 *string) (*gauntlet.Xtruct, error) {
 	structThing := &gauntlet.Xtruct{StringThing: arg1}
 	switch *arg0 {
 	case "Xception":
 		code := int32(1001)
 		message := "This is an Xception"
-		return nil, nil, &gauntlet.Xception{ErrorCode: &code, Message: &message}
+		return nil, &gauntlet.Xception{ErrorCode: &code, Message: &message}
 	case "Xception2":
 		code := int32(2002)
-		return nil, nil, &gauntlet.Xception2{
+		return nil, &gauntlet.Xception2{
 			ErrorCode:   &code,
 			StructThing: structThing,
 		}
 	default:
-		return structThing, resMetaFromReqMeta(reqMeta), nil
+		copyRequestHeaders(ctx)
+		return structThing, nil
 	}
 }
 
-func (thriftTest) TestNest(ctx context.Context, reqMeta yarpc.ReqMeta, thing *gauntlet.Xtruct2) (*gauntlet.Xtruct2, yarpc.ResMeta, error) {
-	return thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestNest(ctx context.Context, thing *gauntlet.Xtruct2) (*gauntlet.Xtruct2, error) {
+	copyRequestHeaders(ctx)
+	return thing, nil
 }
 
-func (thriftTest) TestOneway(ctx context.Context, reqMeta yarpc.ReqMeta, seconds *int32) error {
+func (thriftTest) TestOneway(ctx context.Context, seconds *int32) error {
 	time.Sleep(time.Duration(*seconds) * time.Second)
 	return nil
 }
 
-func (thriftTest) TestSet(ctx context.Context, reqMeta yarpc.ReqMeta, thing map[int32]struct{}) (map[int32]struct{}, yarpc.ResMeta, error) {
-	return thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestSet(ctx context.Context, thing map[int32]struct{}) (map[int32]struct{}, error) {
+	copyRequestHeaders(ctx)
+	return thing, nil
 }
 
-func (thriftTest) TestString(ctx context.Context, reqMeta yarpc.ReqMeta, thing *string) (string, yarpc.ResMeta, error) {
-	return *thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestString(ctx context.Context, thing *string) (string, error) {
+	copyRequestHeaders(ctx)
+	return *thing, nil
 }
 
-func (thriftTest) TestStringMap(ctx context.Context, reqMeta yarpc.ReqMeta, thing map[string]string) (map[string]string, yarpc.ResMeta, error) {
-	return thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestStringMap(ctx context.Context, thing map[string]string) (map[string]string, error) {
+	copyRequestHeaders(ctx)
+	return thing, nil
 }
 
-func (thriftTest) TestStruct(ctx context.Context, reqMeta yarpc.ReqMeta, thing *gauntlet.Xtruct) (*gauntlet.Xtruct, yarpc.ResMeta, error) {
-	return thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestStruct(ctx context.Context, thing *gauntlet.Xtruct) (*gauntlet.Xtruct, error) {
+	copyRequestHeaders(ctx)
+	return thing, nil
 }
 
-func (thriftTest) TestTypedef(ctx context.Context, reqMeta yarpc.ReqMeta, thing *gauntlet.UserId) (gauntlet.UserId, yarpc.ResMeta, error) {
-	return *thing, resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestTypedef(ctx context.Context, thing *gauntlet.UserId) (gauntlet.UserId, error) {
+	copyRequestHeaders(ctx)
+	return *thing, nil
 }
 
-func (thriftTest) TestVoid(ctx context.Context, reqMeta yarpc.ReqMeta) (yarpc.ResMeta, error) {
-	return resMetaFromReqMeta(reqMeta), nil
+func (thriftTest) TestVoid(ctx context.Context) error {
+	copyRequestHeaders(ctx)
+	return nil
 }