+++ internal/examples/thrift-oneway/main.go
@@ -81,7 +81,7 @@ func main() {
 	// Make outbound call every 500ms
 	for {
 		time.Sleep(time.Second / 2)
-		if _, err := client.Sink(context.Background(), nil, &sink.SinkRequest{Message: "hello!"}); err != nil {
+		if _, err := client.Sink(context.Background(), &sink.SinkRequest{Message: "hello!"}); err != nil {
 			log.Print(err)
 		}
 	}
@@ -90,7 +90,7 @@ func main() {
 type helloHandler struct{}
 
 // Sink is our server-side handler implementation
-func (h *helloHandler) Sink(ctx context.Context, reqMeta yarpc.ReqMeta, snk *sink.SinkRequest) error {
+func (h *helloHandler) Sink(ctx context.Context, snk *sink.SinkRequest) error {
 	log.Println("received message:", snk.Message)
 	return nil
 }